+++ lib/bottles.rb
@@ -1,60 +1,35 @@
-require 'pry'
 class Bottles
 
-  def verse(starting_bottles_count)
-    verse = print_first_verse_line(starting_bottles_count)
-    starting_bottles_count -= 1
-    verse += print_second_verse_line(starting_bottles_count)
-    return verse
-  end
-
-  def verses(starting_bottles_count, ending_bottles_count)
-    verses = ""
-    while starting_bottles_count > ending_bottles_count - 1
-      verses += verse(starting_bottles_count)
-      verses += "\n" if starting_bottles_count > ending_bottles_count
-      starting_bottles_count -= 1
-    end
-    return verses
-  end
-
   def song
-    return verses(99, 0)
+    verses(99, 0)
   end
 
-  private
-
-  def pluralize_bottle(bottles_count)
-    if bottles_count == 1
-      return "#{(bottles_count)} bottle"
-    elsif bottles_count == 0
-      return "no more bottles"
-    else
-      return "#{(bottles_count)} bottles"
-    end
+  def verses(starting, ending)
+    starting.downto(ending).map {|i| verse(i)}.join("\n")
   end
 
-  def singularize_take(bottles_count)
-    if bottles_count == 0
-      return "Take it"
+  def verse(number)
+    case number
+    when 0
+      "No more bottles of beer on the wall, " +
+      "no more bottles of beer.\n" +
+      "Go to the store and buy some more, " +
+      "99 bottles of beer on the wall.\n"
+    when 1
+      "1 bottle of beer on the wall, " +
+      "1 bottle of beer.\n" +
+      "Take it down and pass it around, " +
+      "no more bottles of beer on the wall.\n"
+    when 2
+      "2 bottles of beer on the wall, " +
+      "2 bottles of beer.\n" +
+      "Take one down and pass it around, " +
+      "1 bottle of beer on the wall.\n"
     else
-      return "Take one"
-    end
-  end
-
-  def print_first_verse_line(bottles_count)
-    if bottles_count > 0
-      return "#{pluralize_bottle(bottles_count)} of beer on the wall, #{pluralize_bottle(bottles_count)} of beer.\n"
-    else
-      return "No more bottles of beer on the wall, no more bottles of beer.\n"
-    end
-  end
-
-  def print_second_verse_line(bottles_count)
-    if bottles_count > -1
-      return "#{singularize_take(bottles_count)} down and pass it around, #{pluralize_bottle(bottles_count)} of beer on the wall.\n"
-    else
-      return "Go to the store and buy some more, 99 bottles of beer on the wall.\n"
+      "#{number} bottles of beer on the wall, " +
+      "#{number} bottles of beer.\n" +
+      "Take one down and pass it around, " +
+      "#{number-1} bottles of beer on the wall.\n"
     end
   end
 end
